fix : page Components 교체
