+++ index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ko-KR">
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
@@ -9,7 +9,7 @@
       src="https://static.nid.naver.com/js/naveridlogin_js_sdk_2.0.0.js"
       charset="utf-8"
     ></script>
-    <title>Vite + React + TS</title>
+    <title>Apple Farm Store</title>
   </head>
   <body>
     <div id="root"></div>
