+++ src/api/auth/product.ts
@@ -2,5 +2,5 @@ import { instance } from '../auth/axiosInstance';
 
 export const getProductDetail = async (id: number) => {
   const response = await instance.get(`/products/${id}`);
-  return response;
+  return response.data;
 };
